added option to check file count from bt

+++ Friend/firmware/testing/get_info_list.py
@@ -21,7 +21,7 @@ async def main():
 
             while(True):
                  await asyncio.sleep(2.0)
-                 print(np.frombuffer(r,dtype=np.uint8))
+                 print(np.frombuffer(r,dtype=np.uint32))
 
         
 if __name__ == "__main__":
